chore: added typescript contents to js files as preliminary step
