+++ packages/graphiql/src/new-components/Toolbar/Content.js
@@ -1,7 +1,13 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui';
+import { DetailedHTMLProps } from 'react';
 
-const Content = ({ ...props }) => (
+export type ContentProps = DetailedHTMLProps<
+  React.HTMLAttributes<HTMLDivElement>,
+  HTMLDivElement
+>;
+
+const Content = ({ ...props }: ContentProps) => (
   <div
     {...props}
     sx={{
